# server/swagger.yaml
openapi: 3.0.3
info:
  title: Capstone-Express-ORM - OpenAPI 3.0
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:8080/api
...
paths:
...
  /comment:
    post:
      tags:
        - Comments
      security:
        - bearerAuth: []
      summary: create comment image
      operationId: createComment
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                hinhId: 
                  type: string
                noiDung: 
                  type: string          
      responses:
        201:
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        500:
          description: error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    error:
      type: object
      properties:
        error:
          type: string